Refactor: move children split logic to children_isolator_union_split.rs
