+++ src/distributed_planner/mod.rs
@@ -1,4 +1,5 @@
 mod batch_coalescing_below_network_boundaries;
+mod children_isolator_union_split;
 mod distributed_config;
 mod distributed_physical_optimizer_rule;
 mod insert_broadcast;
